♻️  Move SimpleDialog from OldComponents to components
