+++ src/containers/Generation/GenerationDashboard/GenerationAssignmentSection.test.jsx
@@ -3,7 +3,7 @@ import GenerationAssignmentSection from './GenerationAssignmentSection'
 import { render, screen, queryByAttribute, fireEvent, waitFor } from '@testing-library/react'
 import GenerationContext from '../context/GenerationContext'
 import PopUpContext, { PopUpContextProvider } from '../../../context/PopUpContext'
-import SimpleDialog from '../../../components/OldComponents/SimpleDialog'
+import SimpleDialog from '../../../components/SimpleDialog'
 import * as myApi from '../../../services/api'
 import { vi } from 'vitest'
